pyproject.toml
[project]
name = "snowpark-vs-postgres-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Dan Aronson", email = "danaronson@Dans-MacBook-Pro.local" }
]
requires-python = ">=3.11"
dependencies = [
    "psycopg[pool]>=3.2.6",
    "snowflake-connector-python[secure-local-storage]==3.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
